requirements.txt
aioboto3==11.1.0
aiobotocore==2.5.0
aiofauna==0.0.36
aiohttp==3.8.4
aiohttp-cors==0.7.0
aiohttp-devtools==1.0.post0
aiohttp-sse==2.1.0
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.76
botocore==1.29.76
charset-normalizer==3.1.0
click==8.1.3
devtools==0.11.0
dnspython==2.3.0
email-validator==2.0.0.post2
executing==1.2.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
pydantic==1.10.7
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.5.0
urllib3==1.26.15
watchgod==0.8.2
wrapt==1.15.0
yarl==1.9.2
